+++ src/app/sell/page.tsx
@@ -1,9 +1,9 @@
-import NewItemForm from "./NewItemForm";
+import ListingForm from "./ListingForm";
 
-export default function SellAnItemPage() {
+export default function SellPage() {
     return (
-        <main>
-            <NewItemForm />
+        <main className="h-[90vh] flex justify-center items-center">
+            <ListingForm />
         </main>
     );
 }
