+++ src/components/FileInput.tsx
@@ -1,9 +1,9 @@
 
 export default function FileInput() {
     return (
-        <div className="flex items-center justify-center">
+        <div className="flex items-center justify-center size-120">
             <label htmlFor="dropzone-file"
-                   className="flex flex-col items-center justify-center w-full h-64 border-2 border-gray-300 border-dashed rounded-lg cursor-pointer bg-gray-50 ">
+                   className="flex flex-col items-center justify-center w-full h-full border-2 border-gray-300 border-dashed rounded-lg cursor-pointer bg-gray-50">
                 <div className="flex flex-col items-center justify-center pt-5 pb-6">
                     <svg className="w-8 h-8 mb-4 text-gray-500 dark:text-gray-400" aria-hidden="true"
                          xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 16">
